requirements.txt
absl-py==2.3.1
astunparse==1.6.3
azure-core==1.35.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Flask==2.3.3
Flask-Cors==4.0.2
flatbuffers==20181003210633
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msal==1.33.0
msal-extensions==1.3.1
namex==0.1.0
numpy==2.1.3
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
rich==14.1.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tflite-runtime==2.14.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
